GF-8538-defect fixed
In spotlight/Decorators, different decorators are used from different
libraries. But the code change in the previous commit works good only
for the decorators in enyo library. So to generalize the fix across all
libraries, the static 'property' "decorates" is changed to a 'function',
which returns the updated decorator.
Other changes are done according to the comments given by Lex for
previous commit.

Enyo-DCO-1.1-Signed-off-by: Rajyavardhan P <rajyavardhan.p@lge.com>