GF-35832: Address spotlight issues encountered while debugging moon.Drawers oddities:
* Don't prevent dispatching spotlight events if a control is not spottable.
Rationale: It should be the responsibility of the caller to ensure the control is spottable if dispatching something like `onSpotlightFocus` (which it does).

Preventing dispatch is a problem in the following use case: a moon.Button receives an onSpotlightSelect, and in its handler it causes itself to be hidden, and move focus somewhere else (typical, valid use case).  However, since it is hidden, it will never receive the onSpotlightBlur event due to itself (and this causes problems with moon.Button since it needs to change its styling when it's blurred).

An alternative to this change would be to only allow "onSpotlightBlur" events through to non-spottable components, which might be a tad less risky, but I tend to feel like this is the caller's responsibility.

Enyo-DCO-1.1-Signed-Off-By: Kevin Schaaf (kevin.schaaf@lge.com)