GF-58436: Spotlight problem when spotlightPagingControls(the attribute of moon.Scroller) is true.
Problem:
When spotlightPagingControls(the attribute of moon.Scroller) is true and
spotlight is located inside page1, spotlight can move to paging control
area with 5-way mode.But from page2, it can't move.
And when spotlight is located at paging control, it also can't move to
item(in list) with 5-way mode from page2.

Solution:
MoonScrollStrategy is scrolling strategy client by matrix3d but
enyo.Spotlight.Util.getAbsoluteBounds() doen't consider this.
This fix is adding matrix3d x and y value to the left and top of
absolute bounds.