This is a regression of changing spotlight container control from scroller to viewport. https://jira2.lgsvl.com/browse/GF-60891
Issue:
Can not focus on ListActions by pressing down key on close button
In previous, scroller was spotlight 'container'. So, isSpottable returning true by looking it's children.
But, now scroller is not a 'container' but viewport is a 'container'. So, isSpottable returning false.
This is happens when user try to give focus on a control which is not a focusable but it is containing focusable controls in it.

Solution:
I can give focus by spot on scroller because we have the code in spot like below.
Make change on getNearestNeighbor so it can look at it's child controls just like spot did.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com