Revert "Revert "[BHV-15215]Correct control is not spotted when moving out of datagridlist.""
Issue: when spot moves out of any container, the nearest
neighbor
algorithm finds the neighbor based on the container bounds.
That is causing the issue.
Fix:  getNearestNeighbor method is updated to
consider the 'container' scenario and changed the bounds to container to
container's focused child.
Enyo-DCO-1.1-Signed-off-by: Rajyavardhan P rajyavardhan.p@lge.com