Add 'next' property to onSpotlightBlur event
In the case where the currently spotted element is blurred as a
result of different element being spotted, include a reference to
the element about to be spotted in the event payload. This info is
sometimes useful for determining what to do in the blur handler.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)