ENYO-1098: Use simulated flag to distinguish generated tap event
Issue:
When (0, 0) is tapped on drawers, drawers do not open/close.
We have a code to filter out simulated click event in Spotlight onClick handler.
And, it is checking clientX/clientY == 0 to know is that a simulated event.
That is eventually filter out normal click event also.

Fix:
Add a flag to indicate the event is generated from click event.
This can be used in Spotlight to exactly distinguish simulated event.

DCO-1.1-Signed-Off-By: Kunmyon Choi kunmyon.choi@lge.com