Keep track of 5-way selection target
When the user presses Enter, we begin a hold gesture on keydown;
on keyup, we end the hold and perform a Spotlight select.

However, it's possible that the target of the user's action (the
currently spotted element) may disappear between keydown and keyup,
either as a result of the action or coincidentally. If that
happens, we should end the hold gesture as soon as the target
disappears, and we should suppress selection when keyup occurs.

This issue was discovered while reworking the behavior of
Moonstone's paging controls for the 2.6 release.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)