ENYO-1444: Add `spotlightRememberFocus` feature for containers
Historically, a Spotlight container has always restored focus to
its last-spotted child when focus was returned to the container via
a 5-way event.

In some cases, though, it would be preferable to let Spotlight's
nearest-neighbor algorithm find and focus the container child
closest to the 5-way event's originating control. This would be
useful in cases (like a scroller) where we want to keep focus
inside the container by default, but where it's not important (and
may be confusing) to restore focus state when the container is
re-entered.

In this change, we implement a simple version of this feature. If
a container has the property `spotlightRememberFocus: false`, then
we let the nearest-neighbor algorithm find the child to spot when
the container gets 5-way focus.

This simple implementation should be useful in many cases, but in
some cases it may not yield the desired results. Specifically, you
might expect the children of a `spotlightRememberFocus: false`
container to be 'promoted' by the nearest-neighbor algorithm to be
seen as peers to the siblings of the container itself. This
implementation doesn't do that, so a child of the container will
only get focus if the nearest-neighbor algorithm first selects the
container from among its siblings. It would have required much more
significant (and therefore riskier) changes to promote the
container's children, so I opted not to do that for now.

A similar case where you may see surprising results is when
a `spotlightRememberFocus: false` container is nested within
another container that does remember its last focused child. In
this case, the outer container will get to handle the 5-way event
first and may re-spot a different (more recently focused) child,
even if a child of the `spotlightRememberFocus: false` container
appears to be the nearest neighbor of the originating control.

I suspect that we'll want to significantly rework Spotlight in the
not-too-distant future, at which point we can potentially provide a
more complete implementation of this feature.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)