ENYO-1461: Add `preventTap` support
By default, Spotlight fires a synthetic `ontap` event (in
addition to an `onSpotlightSelect` event) when an element is
selected in 5-way mode. This is because Spotlight selection is
semantically equivalent to tapping; controls should generally
take the same action when Spotlight-selected as they do when
tapped.

Library and app developers can therefore generally handle both
cases with a single tap handler, rather than separate handlers
for `ontap` and `onSpotlightSelect`.

However, until now, Spotlight has not respected enyo/gesture's
`preventTap()` method, which allows library or app code to
prevent a tap from occurring as needed. `preventTap()`
is exposed on `up` and `down` events.

In this change, we modify Spotlight to respect `preventTap()`.
We do so by adding the `preventTap()` method to the
`onSpotlightKeyUp` and `onSpotlightKeyDown` events, and
checking to see whether `preventTap()` has been invoked before
we fire the synthetic `ontap` event.

I opted not to prevent the `onSpotlightSelect` event itself, but
just to prevent the `ontap` event. If we encounter a situation
where this is a problem, we can consider preventing both events
in response to the invocation of `preventTap()` or adding an
equivalent `preventSpotlightSelect()` method.

This change was motivated by a use case in
moonstone/NewPagingControl, where we wanted to prevent a tap from
occurring at the end of a hold gesture.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)