ENYO-1664: Revamp spotlight/spot() to provide more/better info...
...about the nature of the `onSpotlightFocus` event.

Spotlight can focus things in response to pointer events, 5-way
key events, or explicit requests to focus a particular Control. It
also sometimes focuses a "default" Control -- for example, when an
app is initialized, when the pointer disappears, or when an app
window regains focus.

Up to now, it has been difficult for app / library code to detect
the nature of a given `onSpotlightFocus` event. Usually it doesn't
matter, but there are some cases where it does, and apps /
libraries have had to resort to some brittle trickery in those
cases.

To improve matters, we introduce the notion of `focusType`, with
possible values being `point`, `5-way`, `explicit` and `default`.
We add `focusType` to the payload of `onSpotlightFocus` so that app
or library code can discriminate as needed. We also include
`direction` in the case of `5-way` events (was previously included
as `dir`, which we leave in for legacy support).

To implement, we changed the API of the `spot()` method to take a
new, general-purpose `info` object as an optional second parameter,
replacing the legacy second and third parameters (though we include
logic to handle those legacy parameters, for backward
compatibility). We also updated internal calls to `spot()`
accordingly.

Related: In spotlight/container, we were previously not
propagating the direction of a 5-way event when it passed through
a container to one of its children. We now do so, so that consumers
of `onSpotlightFocus` can rely on receiving direction info for all
5-way focus events.

These changes were motivated by a case where a scroller would
mysteriously scroll to a different position when the pointer
disappeared or the window regained focus. The mystery scroll
happened because Spotlight was restoring focus to a previously
focused element which was no longer in view. To fix, the scroller
needed to be able to discriminate such "default" focus events from
other types and override the behavior.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)