ENYO-1664: Support "directionless" nearest-neighbor calculations...
...and prepare for a refactoring of nearest neghbor functionality.

While addressing an issue in moonstone/NewDataList where the list
would mysteriously scroll to a different position when the pointer
disappeared or the window regained focus, we experimented with
various fixes. Some of these required the ability to find the
nearest neighbor to an arbitrary point (e.g. the center of the
screen or the center of a scrollable region), without respect to
a specific direction.

While we ultimately ended up choosing a different near-term fix
for ENYO-1664, we leave support for "directionless" nearest-
neighbor calculations in place and make a note suggesting future
enhancements.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)