Add `previous` property to `onSpotlightFocus`, `onSpotlightFocused`
We (fairly) recently added a `next` property to the
`onSpotlightBlur` event, indicating which element will be focused
next.

We now add a similar `previous` property to `onSpotlightFocus` and
`onSpotlightFocused`.

These properties are intended to let handlers discriminate on the
basis of where focus is coming from / going when implementing
special-case logic.

The `previous` property specifically was useful in refining 5-way
scrolling behavior in Moonstone, to switch between item-at-a-time
and page-at-a-time scrolling depending on where the focus
previously was (see moonstone/Scrollable).

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)