blur the activeElement when spotting a non-focusable control
If spotlight alternates between a spottable non-label and a label or
control with accessibility disabled, focus isn't changed when returning
to the non-label because it wasn't set to the label when it was
spotted. So, if the spotted control can't receive focus and there is a
document.activeElement, blur it so it can receive focus again.

Issue: ENYO-2431
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)