ensure only 1 disappearance timeout is active
The setTimeout callback was using variables whose values could be no
longer valid when the callback was executed. oExCurrent would always
point to the previous at the time the callback was called but may not
be the previous at the time the timeout was scheduled. Similarly,
oControl may not need to be observed if it was no longer the currently
spotted control when the callback is called.

Instead, ensure that the timeout is only scheduled if there isn't one
pending, bind oExCurrent to the callback so the previously spotted
control is correctly unobserved, and use _oCurrent in the callback so
the currently spotted control is observed.

Issue: ENYO-2756
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)