ENYO-2744: Allow setting lastFocusedChild to null
We have long had, for special cases, a public API allowing the
last focused child of a Spotlight container to be explicitly set.

We have a new use case where we want to set it to `null`, so
extending the API accordingly.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)