allow onSpotlightContainerEnter event when moving from self to child
Previously, the isDescendantOf check prevented the event when the
spotted control was a child of a container that was the previously
spotted control (_oLastControl).

This could occur when the first spottable child in the app was a
spotlight container (e.g. a Scroller or a Panel) and Spotlight was
in pointer mode during init. In this scenario, hitting a 5-way key
after startup would move focus from the container to one of its
children but it wouldn't receive the onSpotlightContainerEnter event
because it moved from itself to a child. This fix requires that the
event pass through if the previously spotted control (oEvent.last) is
the sender of the event (the container).

Issue: ENYO-2796
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)