ENYO-3091: Support spotlightRememberFocus option on setLastFocusedChild
Issue:
We have spotlightRememberFocus option on spotlight container, but this only work on 5way key press case. So, if we want to apply this option on popup then it is not work in expected way. Meaning, remember last focus on show.

Fix:
Add a condition on setLastFocusedChild to not remember last focused control when spotlightRememberFocus is false.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi kunmyon.choi@lge.com