ENYO-3421: Spotlight focus jump or lost when show/hide popup repeatedly
Issue:
When open and close popup by 5way, we expect that we are in 5way mode.
Unexpected mouse move event with same position fired if we do dom.focus
on animating element.

We have a logic that is dropping mouse move event which doesn't
actually changed position. Logically, it should be filtered out in here.
But, problem happens when pointer mode is false because it is only update
last position on pointer mode.

Fix:
We remove spotlight pointer mode check from clientXYChanged so that it
can always cache last mouse position to compare movement.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi (kunmyon.choi@lge.com)