ENYO-3451: Dispatch onSpotlightBlur on disappear
Issue:
We dont dispatch onSpotlightBlur on onDisappear routine. That is because
we doesn't have any usecases for this case. But, store demo app recently
reported an issue regarding to this because they need to recover some
state on blur event.

Fix:
We dispatch onSpotlightBlur on disappear, so that any control that is
unhighlight focus can get event.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi (kunmyon.choi@lge.com)